MMORPG_bulletin_board/requirements.txt
cffi==1.15.0
defusedxml==0.7.1
Pillow==10.1.0
ipython==8.0.0
pip==21.2.4
tornado==6.1
numpy==1.22.4
setuptools==58.1.0
Django==4.2.7
asgiref==3.7.2
sqlparse==0.4.4
Jinja2==3.0.3
requests==2.27.1
django-filter==23.3
tzdata==2023.3